fix(k8s): resolve mock mismatches, add shared mocks, and prevent SSH fallback in tests
