+++ internal/repositories/k8s/kubeadm_provisioner_test.go
@@ -43,7 +43,11 @@ func (m *MockInstanceService) StopInstance(ctx context.Context, id string) error
 	return nil
 }
 func (m *MockInstanceService) ListInstances(ctx context.Context) ([]*domain.Instance, error) {
-	return nil, nil
+	args := m.Called(ctx)
+	if args.Get(0) == nil {
+		return nil, args.Error(1)
+	}
+	return args.Get(0).([]*domain.Instance), args.Error(1)
 }
 func (m *MockInstanceService) GetInstanceLogs(ctx context.Context, id string) (string, error) {
 	return "", nil
@@ -325,7 +329,15 @@ kubeadm join 10.0.0.100:6443 --token abc --discovery-token-ca-cert-hash sha256:1
 
 kubeadm join 10.0.0.100:6443 --token abc --discovery-token-ca-cert-hash sha256:123 --control-plane --certificate-key xyz
 `
+	// Collect instances for ListInstances mock
+	var instances []*domain.Instance
+	for _, n := range allNodes {
+		inst, _ := instSvc.GetInstance(ctx, n.InstanceID.String())
+		instances = append(instances, inst)
+	}
+
 	instSvc.On("Exec", mock.Anything, mock.Anything, mock.Anything).Return(kubeadmOutput, nil).Maybe()
+	instSvc.On("ListInstances", mock.Anything).Return(instances, nil).Maybe()
 
 	err := p.Provision(ctx, cluster)
